updated_without_event_bus
