+++ src/consts/index.ts
@@ -99,6 +99,14 @@ export const LOCAL_STORAGE_MAX_SIZE = 4 * 1024 * 1024;
 
 export const LOCAL_STORAGE_LIMIT_PERCENTAGE = 95;
 
+export const listOfRemoveForLocalStorage = [
+  '.assetsAddresses',
+  '.history',
+  '.ethBridgeHistory',
+  '.evmHistory',
+  '.subHistory',
+];
+
 export enum MarketAlgorithms {
   SMART = 'SMART',
   TBC = 'TBC',
