--- src/eventBus.ts
@@ -1,2 +1,0 @@
-import Vue from 'vue';
-export const EventBus = new Vue();
